Backend/requirements.txt
django==3.2
django-cors-headers
djangorestframework
mysqlclient